feat: update UIBackgroundModes and continuous reproduction on foreground and pertinant notification.
