+++ src/app/Home.tsx
@@ -20,28 +20,8 @@ import {
 export const Home = () => {
   const [key, setKey] = useState(0);
 
-  const appState = useRef(AppState.currentState);
-
   const { isHome, isAboutUs, isContact } = useAppContext();
 
-  useEffect(() => {
-    const subscription = AppState.addEventListener("change", (nextAppState) => {
-      if (
-        appState.current.match(/inactive|background/) &&
-        nextAppState === "active"
-      ) {
-        console.log("App has come to the foreground!");
-      }
-
-      appState.current = nextAppState;
-      setKey((prevKey) => prevKey + 1);
-      console.log("AppState", appState.current);
-    });
-
-    return () => {
-      subscription.remove();
-    };
-  }, []);
 
   useEffect(() => {
     const unsubscribe = NetInfo.addEventListener((state) => {
